+++ src/test/scala/ai/tripl/arc/transform/typing/LongTypingSuite.scala
@@ -287,5 +287,34 @@ class LongTypingSuite extends FunSuite with BeforeAndAfter {
         }
       }
     }
+
+    //test scientific number
+    {
+      val col = LongColumn(None, name = "name", description = Some("description"), nullable = false, nullReplacementValue = None, trim = false, nullableValues = "" :: Nil, metadata=None, formatters = None)
+
+      // value contains E
+      {
+        val value = "7.987654321000000E+12"
+        Typing.typeValue(value, col) match {
+          case (Some(res), err) => {
+            assert(res === 7987654321000L)
+            assert(err === None)
+          }
+          case (_, _) => assert(false)
+        }
+      }
+
+      {
+        val value = "9.87650000000000000e+14"
+        Typing.typeValue(value, col) match {
+          case (Some(res), err) => {
+            assert(res === 987650000000000L)
+            assert(err === None)
+          }
+          case (_, _) => assert(false)
+        }
+      }
+    }
+
   }
 }